backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
argon2-cffi==23.1.0
python-multipart==0.0.19
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0
